% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inftheo.R
\name{check_type}
\alias{check_type}
\title{Internal function to check the variable's type.  Information theory
functions may not work as intended on doubles, so make a warning.
Sometimes doubles are the unexpected column type though.}
\usage{
check_type(tab, varname)
}
\value{
the table with the column possibly coerced as a character
}
\description{
Internal function to check the variable's type.  Information theory
functions may not work as intended on doubles, so make a warning.
Sometimes doubles are the unexpected column type though.
}
\keyword{internal}
