% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{mutual_info_heatmap}
\alias{mutual_info_heatmap}
\title{Plot heatmap of mutual infos}
\usage{
mutual_info_heatmap(mi_matrix, xlabel = NULL, ylabel = NULL, title = NULL,
  font_sizes = c(12, 12))
}
\arguments{
\item{mi_matrix}{a table produced from \link{mutual_info_matrix}}

\item{xlabel}{x-axis label}

\item{ylabel}{y-axis label}

\item{title}{title of plot}

\item{font_sizes}{A length-2 vector of x-axis and y-axis variable font sizes}
}
\value{
a double with the calculated value
}
\description{
Given a matrix produced from mutual_info_matrix(), plot a heatmap with ggplot2
}
\examples{
p <- mtcars \%>\%
   mutual_info_matrix(cyl, vs, am, gear, carb, normalized=TRUE) \%>\%
   mutual_info_heatmap()
}
\seealso{
\link{mutual_info_matrix}
}
