% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutual_info.R
\name{shannon_cond_entropy}
\alias{shannon_cond_entropy}
\title{Conditional Shannon Entropy H(X|Y) i.e. "H(X given Y)"}
\usage{
shannon_cond_entropy(tab, ...)
}
\arguments{
\item{tab}{A tibble with the columns of interest}

\item{...}{two columns (variables) selected}
}
\value{
a double with the calculated value
}
\description{
This calculates conditional shannon entropy of two columns in a tibble,
according to:
$$H(X|Y) = -\sum_{y\in{Y}}\sum_{x\in{X}}p(x|y)f(p(x|y))$$
It's assumed these columns are character typed with no NAs.
}
\seealso{
\code{\link[=shannon_entropy]{shannon_entropy()}}
}
