% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inftheo.R
\name{shannon_entropy}
\alias{shannon_entropy}
\title{Shannon Entropy H(X)}
\usage{
shannon_entropy(.data, X, na.rm = FALSE)
}
\arguments{
\item{.data}{A tibble with the column of interest}

\item{X}{Name of the column}

\item{na.rm}{remove all rows with NA values in at least one of the columns}
}
\value{
a double with the calculated value
}
\description{
This calculates shannon entropy of a variable in a tibble.
It's assumed these columns are character typed with no NAs.
}
\examples{
shannon_entropy(iris, Species)
iris \%>\% as_tibble() \%>\% shannon_entropy(Species)
shannon_entropy(iris, 'Species')
shannon_entropy(iris, 5)
}
\seealso{
\link{shannon_cond_entropy}
}
