% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infotheo.R
\docType{package}
\name{tidyinftheo-package}
\alias{tidyinftheo}
\alias{tidyinftheo-package}
\title{tidyinfotheo: tidy-style information theoretic routines}
\description{
tidyinfotheo provides an easy way to calculate shannon entropy,
mutual information, etc. from fields in a 'tibble'
}
\details{
To learn more about tidyinfotheo, start with the vignettes:
\code{browseVignettes(package = "tidyinfotheo")}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pohlio/tidyinfotheo}
  \item Report bugs at \url{https://github.com/pohlio/tidyinfotheo/issues}
}

}
\author{
\strong{Maintainer}: Andy Pohl \email{andy@pohl.io}

Other contributors:
\itemize{
  \item Pohl Data, LLC [copyright holder]
}

}
