% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infotheo.R
\name{mutual_info}
\alias{mutual_info}
\title{Mutual information MI(X;Y) = H(X) - H(X|Y) = H(Y) - H(Y|X)}
\usage{
mutual_info(tab, ..., normalized = FALSE, na.rm = FALSE)
}
\arguments{
\item{tab}{A tibble with the column of interest}

\item{...}{two columns (variables) selected}

\item{normalized}{if TRUE, scale to \link{0,1}}
}
\value{
a double with the calculated value
}
\description{
This calculates the mutual information between two variables in a tibble.
(if normalized).  It's assumed these columns are character typed with no NAs.
}
