% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inftheo.R
\name{shannon_cond_entropy}
\alias{shannon_cond_entropy}
\title{Conditional Shannon Entropy H(X|Y) i.e. "H(X given Y)"}
\usage{
shannon_cond_entropy(tab, ..., na.rm = FALSE)
}
\arguments{
\item{tab}{A tibble with the columns of interest}

\item{...}{two columns (variables) selected}
}
\value{
a double with the calculated value
}
\description{
This calculates conditional shannon entropy of two columns in a tibble.
It's assumed these columns are character typed with no NAs.
}
\seealso{
\link{shannon_entropy}
}
