% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutual_info.R
\name{shannon_entropy}
\alias{shannon_entropy}
\title{Shannon Entropy H(X)}
\usage{
shannon_entropy(tab, X)
}
\arguments{
\item{tab}{A tibble with the column of interest}

\item{X}{Name of the column}
}
\value{
a double with the calculated value
}
\description{
This calculates shannon entropy of a variable in a tibble,
according to:
$$H(X) = -\sum_{x\in{X}}p(x)f(x)$$
It's assumed these columns are character typed with no NAs.
}
\seealso{
\code{\link[=shannon_cond_entropy]{shannon_cond_entropy()}}
}
