% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inftheo.R
\docType{package}
\name{tidyinftheo-package}
\alias{tidyinftheo}
\alias{tidyinftheo-package}
\title{tidyinftheo: tidy-style information theoretic routines}
\description{
tidyinftheo provides an easy way to calculate shannon entropy,
mutual information, etc. from fields in a 'tibble'
}
\details{
To learn more about tidyinftheo, start with the vignettes:
\code{browseVignettes(package = "tidyinftheo")}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pohlio/tidyinftheo}
  \item Report bugs at \url{https://github.com/pohlio/tidyinftheo/issues}
}

}
\author{
\strong{Maintainer}: Andy Pohl \email{andy@pohl.io}

Other contributors:
\itemize{
  \item Pohl Data, LLC [copyright holder]
}

}
